enforce strong return typing

+++ modules/thermo/common/src/main/kotlin/com/mussonindustrial/embr/thermo/scripting/IF97ScriptModuleImpl.kt
@@ -29,12 +29,12 @@ class IF97ScriptModuleImpl : IF97ScriptModule {
         types = [Double::class, Double::class, Double::class, Double::class],
     )
     override fun compressibility(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.compressibility.call(args, keywords) as Double
+        IF97PyArgOverloads.compressibility.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(names = ["p", "t"], types = [Double::class, Double::class])
     override fun compressionFactor(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.compressionFactor.call(args, keywords) as Double
+        IF97PyArgOverloads.compressionFactor.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(
@@ -42,7 +42,7 @@ class IF97ScriptModuleImpl : IF97ScriptModule {
         types = [Double::class, Double::class, Double::class, Double::class, Double::class],
     )
     override fun density(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.density.call(args, keywords) as Double
+        IF97PyArgOverloads.density.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(
@@ -50,7 +50,7 @@ class IF97ScriptModuleImpl : IF97ScriptModule {
         types = [Double::class, Double::class, Double::class, Double::class, Double::class],
     )
     override fun dielectricConstant(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.dielectricConstant.call(args, keywords) as Double
+        IF97PyArgOverloads.dielectricConstant.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(
@@ -58,7 +58,7 @@ class IF97ScriptModuleImpl : IF97ScriptModule {
         types = [Double::class, Double::class, Double::class, Double::class],
     )
     override fun heatCapacityRatio(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.heatCapacityRatio.call(args, keywords) as Double
+        IF97PyArgOverloads.heatCapacityRatio.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(
@@ -66,7 +66,7 @@ class IF97ScriptModuleImpl : IF97ScriptModule {
         types = [Double::class, Double::class, Double::class, Double::class],
     )
     override fun isentropicExponent(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.isentropicExponent.call(args, keywords) as Double
+        IF97PyArgOverloads.isentropicExponent.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(
@@ -76,7 +76,7 @@ class IF97ScriptModuleImpl : IF97ScriptModule {
     override fun isobaricCubicExpansionCoefficient(
         args: Array<PyObject>,
         keywords: Array<String>,
-    ): Double = IF97PyArgOverloads.isobaricCubicExpansionCoefficient.call(args, keywords) as Double
+    ): Double = IF97PyArgOverloads.isobaricCubicExpansionCoefficient.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(
@@ -84,7 +84,7 @@ class IF97ScriptModuleImpl : IF97ScriptModule {
         types = [Double::class, Double::class, Double::class, Double::class],
     )
     override fun isobaricHeatCapacity(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.isobaricHeatCapacity.call(args, keywords) as Double
+        IF97PyArgOverloads.isobaricHeatCapacity.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(
@@ -92,7 +92,7 @@ class IF97ScriptModuleImpl : IF97ScriptModule {
         types = [Double::class, Double::class, Double::class, Double::class],
     )
     override fun isochoricHeatCapacity(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.isochoricHeatCapacity.call(args, keywords) as Double
+        IF97PyArgOverloads.isochoricHeatCapacity.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(
@@ -100,7 +100,7 @@ class IF97ScriptModuleImpl : IF97ScriptModule {
         types = [Double::class, Double::class, Double::class, Double::class, Double::class],
     )
     override fun kinematicViscosity(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.kinematicViscosity.call(args, keywords) as Double
+        IF97PyArgOverloads.kinematicViscosity.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(
@@ -108,12 +108,12 @@ class IF97ScriptModuleImpl : IF97ScriptModule {
         types = [Double::class, Double::class, Double::class, Double::class],
     )
     override fun prandtl(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.prandtl.call(args, keywords) as Double
+        IF97PyArgOverloads.prandtl.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(names = ["h", "s"], types = [Double::class, Double::class])
     override fun pressure(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.pressure.call(args, keywords) as Double
+        IF97PyArgOverloads.pressure.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(
@@ -129,17 +129,17 @@ class IF97ScriptModuleImpl : IF97ScriptModule {
             ],
     )
     override fun refractiveIndex(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.refractiveIndex.call(args, keywords) as Double
+        IF97PyArgOverloads.refractiveIndex.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(names = ["h", "s", "t"], types = [Double::class, Double::class, Double::class])
     override fun saturationPressure(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.saturationPressure.call(args, keywords) as Double
+        IF97PyArgOverloads.saturationPressure.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(names = ["h", "p", "s"], types = [Double::class, Double::class, Double::class])
     override fun saturationTemperature(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.saturationTemperature.call(args, keywords) as Double
+        IF97PyArgOverloads.saturationTemperature.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(
@@ -147,21 +147,21 @@ class IF97ScriptModuleImpl : IF97ScriptModule {
         types = [Double::class, Double::class, Double::class, Double::class],
     )
     override fun specificEnthalpy(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.specificEnthalpy.call(args, keywords) as Double
+        IF97PyArgOverloads.specificEnthalpy.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(names = ["p", "t"], types = [Double::class, Double::class])
     override fun specificEnthalpySaturatedLiquid(
         args: Array<PyObject>,
         keywords: Array<String>,
-    ): Double = IF97PyArgOverloads.specificEnthalpySaturatedLiquid.call(args, keywords) as Double
+    ): Double = IF97PyArgOverloads.specificEnthalpySaturatedLiquid.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(names = ["p", "t"], types = [Double::class, Double::class])
     override fun specificEnthalpySaturatedVapor(
         args: Array<PyObject>,
         keywords: Array<String>,
-    ): Double = IF97PyArgOverloads.specificEnthalpySaturatedVapor.call(args, keywords) as Double
+    ): Double = IF97PyArgOverloads.specificEnthalpySaturatedVapor.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(
@@ -169,26 +169,26 @@ class IF97ScriptModuleImpl : IF97ScriptModule {
         types = [Double::class, Double::class, Double::class, Double::class],
     )
     override fun specificEntropy(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.specificEntropy.call(args, keywords) as Double
+        IF97PyArgOverloads.specificEntropy.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(names = ["p", "t"], types = [Double::class, Double::class])
     override fun specificEntropySaturatedLiquid(
         args: Array<PyObject>,
         keywords: Array<String>,
-    ): Double = IF97PyArgOverloads.specificEntropySaturatedLiquid.call(args, keywords) as Double
+    ): Double = IF97PyArgOverloads.specificEntropySaturatedLiquid.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(names = ["p", "t"], types = [Double::class, Double::class])
     override fun specificEntropySaturatedVapor(
         args: Array<PyObject>,
         keywords: Array<String>,
-    ): Double = IF97PyArgOverloads.specificEntropySaturatedVapor.call(args, keywords) as Double
+    ): Double = IF97PyArgOverloads.specificEntropySaturatedVapor.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(names = ["p", "t"], types = [Double::class, Double::class])
     override fun specificGibbsFreeEnergy(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.specificGibbsFreeEnergy.call(args, keywords) as Double
+        IF97PyArgOverloads.specificGibbsFreeEnergy.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(
@@ -196,23 +196,21 @@ class IF97ScriptModuleImpl : IF97ScriptModule {
         types = [Double::class, Double::class, Double::class, Double::class, Double::class],
     )
     override fun specificInternalEnergy(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.specificInternalEnergy.call(args, keywords) as Double
+        IF97PyArgOverloads.specificInternalEnergy.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(names = ["p", "t"], types = [Double::class, Double::class])
     override fun specificInternalEnergySaturatedLiquid(
         args: Array<PyObject>,
         keywords: Array<String>,
-    ): Double =
-        IF97PyArgOverloads.specificInternalEnergySaturatedLiquid.call(args, keywords) as Double
+    ): Double = IF97PyArgOverloads.specificInternalEnergySaturatedLiquid.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(names = ["p", "t"], types = [Double::class, Double::class])
     override fun specificInternalEnergySaturatedVapor(
         args: Array<PyObject>,
         keywords: Array<String>,
-    ): Double =
-        IF97PyArgOverloads.specificInternalEnergySaturatedVapor.call(args, keywords) as Double
+    ): Double = IF97PyArgOverloads.specificInternalEnergySaturatedVapor.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(
@@ -220,21 +218,21 @@ class IF97ScriptModuleImpl : IF97ScriptModule {
         types = [Double::class, Double::class, Double::class, Double::class, Double::class],
     )
     override fun specificVolume(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.specificVolume.call(args, keywords) as Double
+        IF97PyArgOverloads.specificVolume.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(names = ["p", "t"], types = [Double::class, Double::class])
     override fun specificVolumeSaturatedLiquid(
         args: Array<PyObject>,
         keywords: Array<String>,
-    ): Double = IF97PyArgOverloads.specificVolumeSaturatedLiquid.call(args, keywords) as Double
+    ): Double = IF97PyArgOverloads.specificVolumeSaturatedLiquid.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(names = ["p", "t"], types = [Double::class, Double::class])
     override fun specificVolumeSaturatedVapor(
         args: Array<PyObject>,
         keywords: Array<String>,
-    ): Double = IF97PyArgOverloads.specificVolumeSaturatedVapor.call(args, keywords) as Double
+    ): Double = IF97PyArgOverloads.specificVolumeSaturatedVapor.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(
@@ -242,17 +240,17 @@ class IF97ScriptModuleImpl : IF97ScriptModule {
         types = [Double::class, Double::class, Double::class, Double::class],
     )
     override fun speedOfSound(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.speedOfSound.call(args, keywords) as Double
+        IF97PyArgOverloads.speedOfSound.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(names = ["p", "t"], types = [Double::class, Double::class])
     override fun surfaceTension(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.surfaceTension.call(args, keywords) as Double
+        IF97PyArgOverloads.surfaceTension.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(names = ["h", "p", "s"], types = [Double::class, Double::class, Double::class])
     override fun temperature(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.temperature.call(args, keywords) as Double
+        IF97PyArgOverloads.temperature.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(
@@ -260,7 +258,7 @@ class IF97ScriptModuleImpl : IF97ScriptModule {
         types = [Double::class, Double::class, Double::class, Double::class, Double::class],
     )
     override fun thermalConductivity(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.thermalConductivity.call(args, keywords) as Double
+        IF97PyArgOverloads.thermalConductivity.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(
@@ -268,7 +266,7 @@ class IF97ScriptModuleImpl : IF97ScriptModule {
         types = [Double::class, Double::class, Double::class, Double::class],
     )
     override fun thermalDiffusivity(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.thermalDiffusivity.call(args, keywords) as Double
+        IF97PyArgOverloads.thermalDiffusivity.call(args, keywords)
 
     @JythonElement(docBundlePrefix = BUNDLE_PREFIX)
     @KeywordArgs(
@@ -276,5 +274,5 @@ class IF97ScriptModuleImpl : IF97ScriptModule {
         types = [Double::class, Double::class, Double::class, Double::class],
     )
     override fun vaporFraction(args: Array<PyObject>, keywords: Array<String>): Double =
-        IF97PyArgOverloads.vaporFraction.call(args, keywords) as Double
+        IF97PyArgOverloads.vaporFraction.call(args, keywords)
 }
